pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pctl"
version = "0.3.0"
description = "PAIC Control - Unified testing CLI for token, journey, and ELK management"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "cryptography>=45.0.6",
    "httpx>=0.28.1",
    "jwcrypto>=1.5.6",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
pctl = "pctl.cli.main:cli"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
    "requests>=2.32.5",
    "ruff>=0.12.11",
]
